requirements.txt
crewai-tools
crewai
langchain_google_genai
python-dotenv